fix: correct voting requesting in fellowship
